Build 3
This build merges static and index into one file, where if fetching to flickr fails, it will fetch some static files instead. This also updates some dependencies.
