+++ package.json
@@ -61,7 +61,8 @@
       "tailwindcss-animate": "^1.0.7",
       "twemoji": "^14.0.2",
       "typescript": "^5.3.3",
-      "vaul": "^0.9.0"
+      "vaul": "^0.9.0",
+      "vercel": "^33.5.3"
    },
    "devDependencies": {
       "@types/bad-words": "^3.0.3",
